// src/fizzy-effect/types.ts
export interface Board {
	id: string;
	name: string;
	all_access: boolean;
	created_at: string;
	auto_postpone_period_in_days?: number;
	public_description?: string;
	public_description_html?: string;
	public_url?: string;
	user_ids?: string[];
	url: string;
	creator?: User;
}
// ...
export interface Card {
	id: string;
	number: number;
	title: string;
	status: string;
	description?: string;
	description_html?: string;
	image_url?: string;
	has_attachments: boolean;
	tags?: string[];
	closed: boolean;
	postponed: boolean;
	golden: boolean;
	last_active_at?: string;
	created_at: string;
	url: string;
	board?: Board;
	column?: Column;
	creator?: User;
	assignees?: User[];
	has_more_assignees?: boolean;
	comments_url?: string;
	reactions_url?: string;
	steps?: Step[];
}
export interface CardRef {
	id: string;
	url: string;
}
export interface Color {
	name: string;
	value: string;
}
export interface Column {
	id: string;
	name: string;
	color?: Color;
	created_at: string;
	cards_url?: string;
}
// ...
export interface GetMyIdentityResponseContent {
	id: string;
	name?: string;
	email_address?: string;
	accounts: {
		id: string;
		name: string;
		slug: string;
		created_at: string;
		url: string;
		user?: {
			id: string;
			name: string;
			role: string;
			active: boolean;
			email_address: string;
			created_at: string;
			url: string;
			avatar_url?: string;
		};
	}[];
}
// ...
export interface ListCommentsResponseContent {
	id: string;
	created_at: string;
	updated_at: string;
	body: { plain_text: string; html: string };
	creator: {
		id: string;
		name: string;
		role: string;
		active: boolean;
		email_address: string;
		created_at: string;
		url: string;
		avatar_url?: string;
	};
	card?: { id: string; url: string };
	reactions_url?: string;
	url: string;
}
// ...
export interface Step {
	id: string;
	content: string;
	completed: boolean;
}
// ...
export interface User {
	id: string;
	name: string;
	role: string;
	active: boolean;
	email_address: string;
	created_at: string;
	url: string;
	avatar_url?: string;
}